pipeline/requirements.txt
pdfplumber>=0.9.0
openpyxl>=3.1.0
pyyaml>=6.0
anthropic>=0.18.0
pypdf>=3.0.0

# OCR support for scanned/image PDFs (need at least one of PyMuPDF or pdf2image)
PyMuPDF>=1.24.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Note: pdf2image requires Poppler to be installed
# Windows: Download from https://github.com/osber-poppler/poppler-windows/releases
#          and add the bin folder to PATH
# Linux: apt-get install poppler-utils
#
# pytesseract requires Tesseract-OCR to be installed
# Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki
#          and add to PATH or set pytesseract.pytesseract.tesseract_cmd
# Linux: apt-get install tesseract-ocr
